refactor: split frontend into modules
