+++ src/web_app/static/chat.ts
@@ -9,6 +9,8 @@ export function setupChat() {
   chatHistory = document.getElementById('chat-history')!;
   chatForm = document.getElementById('chat-form') as HTMLFormElement;
   chatInput = document.getElementById('chat-input') as HTMLInputElement;
+  const closeBtn = chatModal.querySelector('.close') as HTMLElement;
+  closeBtn?.addEventListener('click', closeChat);
 
   chatForm?.addEventListener('submit', async (e) => {
     e.preventDefault();
