+++ src/web_app/static/main.ts
@@ -1,10 +1,13 @@
-import { setupUpload } from './upload';
-import { setupFolders, refreshFolderTree } from './folders';
-import { setupChat } from './chat';
+import { setupUpload } from './upload.js';
+import { setupFiles, refreshFiles } from './files.js';
+import { setupFolders, refreshFolderTree } from './folders.js';
+import { setupChat } from './chat.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   setupUpload();
+  setupFiles();
   setupFolders();
   setupChat();
+  refreshFiles();
   refreshFolderTree();
 });
