+++ src/web_app/templates/index.html
@@ -6,6 +6,7 @@
     <link rel="stylesheet" href="/static/style.css">
     <link href="https://unpkg.com/cropperjs@1.5.13/dist/cropper.min.css" rel="stylesheet" />
     <script src="https://unpkg.com/cropperjs@1.5.13/dist/cropper.min.js" defer></script>
+    <!-- Подключаем собранный модуль -->
     <script type="module" src="/static/dist/main.js"></script>
 </head>
 <body>
