+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "ES6",
     "module": "ES6",
+    "moduleResolution": "bundler",
     "rootDir": "src/web_app/static",
     "outDir": "src/web_app/static/dist",
     "strict": false,
